동영상찍어야해...무사수행은안해..

+++ princessMaker2/inventoryScene.h
@@ -32,7 +32,7 @@ public:
 
 	void setDialog(string dialog);
 
-	void changeStatus(string name, int value);
+	void changeStatus(string name, float value);
 
 	bool dialogRender();
